- Fix a type bug. - Enable OpemMP in debug.

+++ src/libaten/accelerator/sbvh.h
@@ -14,8 +14,8 @@ namespace aten {
 		float miss{ -1 };		///< Link index if ray miss.
 
 #if (SBVH_TRIANGLE_NUM == 1)
-		int triid{ -1 };
-		int parent{ -1 };
+		float triid{ -1 };
+		float parent{ -1 };
 		float padding[2];
 
 		bool isLeaf() const
